+++ storage/archive/archive_performance.c
@@ -40,8 +40,8 @@ char test_string[BUFFER_LEN];
 
 /* prototypes */
 long int timedif(struct timeval a, struct timeval b);
-int generate_data(unsigned long long length);
-int read_test(azio_stream *reader_handle, unsigned long long rows_to_test_for);
+int generate_data(uint64_t length);
+int read_test(azio_stream *reader_handle, uint64_t rows_to_test_for);
 
 int main(int argc, char *argv[])
 {
@@ -89,10 +89,10 @@ int main(int argc, char *argv[])
   return 0;
 }
 
-int generate_data(unsigned long long rows_to_test)
+int generate_data(uint64_t rows_to_test)
 {
   azio_stream writer_handle;
-  unsigned long long x;
+  uint64_t x;
   unsigned int ret;
   struct timeval start_time, end_time;
   long int timing;
@@ -140,10 +140,10 @@ int generate_data(unsigned long long rows_to_test)
   return 0;
 }
 
-int read_test(azio_stream *reader_handle, unsigned long long rows_to_test_for)
+int read_test(azio_stream *reader_handle, uint64_t rows_to_test_for)
 {
-  unsigned long long read_length= 0;
-  unsigned long long count= 0;
+  uint64_t read_length= 0;
+  uint64_t count= 0;
   unsigned int ret;
   int error;
 
@@ -152,7 +152,7 @@ int read_test(azio_stream *reader_handle, unsigned long long rows_to_test_for)
   {
     if (error)
     {
-      fprintf(stderr, "Got an error while reading at row %llu\n", count);
+      fprintf(stderr, "Got an error while reading at row %"PRIu64"\n", count);
       exit(1);
     }
